requirements.txt
Django==1.11.2
dj-database-url==0.4.2
django-extensions==1.7.9
django-grappelli==2.10.1
django-import-export==0.5.1
django-test-without-migrations==0.6
pandas==0.20.1
python-decouple==3.0
